refactor: MessageTemplate 클래스의 패키지 변경
view 에서 util 패키지로 변경
