+++ src/main/java/store/view/InputView.java
@@ -3,8 +3,8 @@ package store.view;
 import static store.util.Constant.ADDITIONAL_PURCHASE_MESSAGE;
 import static store.util.Constant.MEMBERSHIP_DISCOUNT_MESSAGE;
 import static store.util.Constant.ORDER_MESSAGE;
-import static store.view.MessageTemplate.PROMOTION_BENEFIT;
-import static store.view.MessageTemplate.PROMOTION_QUANTITY_SHORTAGE;
+import static store.util.MessageTemplate.PROMOTION_BENEFIT;
+import static store.util.MessageTemplate.PROMOTION_QUANTITY_SHORTAGE;
 
 import camp.nextstep.edu.missionutils.Console;
