+++ src/main/java/store/view/OutputView.java
@@ -6,14 +6,14 @@ import static store.util.ErrorMessage.ERROR_TAG;
 import static store.util.Constant.RECEIPT_AMOUNT_INFO_TITLE;
 import static store.util.Constant.RECEIPT_GIVEAWAY_TITLE;
 import static store.util.Constant.RECEIPT_TITLE;
-import static store.view.MessageTemplate.PRODUCT_INFO;
-import static store.view.MessageTemplate.PRODUCT_INFO_ZERO_QUANTITY;
-import static store.view.MessageTemplate.RECEIPT_DISCOUNT;
-import static store.view.MessageTemplate.RECEIPT_GIVEAWAY_HISTORY;
-import static store.view.MessageTemplate.RECEIPT_HEADER;
-import static store.view.MessageTemplate.RECEIPT_PAYMENT;
-import static store.view.MessageTemplate.RECEIPT_PURCHASE_HISTORY;
-import static store.view.MessageTemplate.RECEIPT_TOTAL_PURCHASE_AMOUNT;
+import static store.util.MessageTemplate.PRODUCT_INFO;
+import static store.util.MessageTemplate.PRODUCT_INFO_ZERO_QUANTITY;
+import static store.util.MessageTemplate.RECEIPT_DISCOUNT;
+import static store.util.MessageTemplate.RECEIPT_GIVEAWAY_HISTORY;
+import static store.util.MessageTemplate.RECEIPT_HEADER;
+import static store.util.MessageTemplate.RECEIPT_PAYMENT;
+import static store.util.MessageTemplate.RECEIPT_PURCHASE_HISTORY;
+import static store.util.MessageTemplate.RECEIPT_TOTAL_PURCHASE_AMOUNT;
 
 import java.util.List;
 import java.util.Map;
